plugins/top10-scan/test-fixtures/vulnerable-fastapi/requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
sqlalchemy==1.4.46
python-jose==3.3.0
passlib==1.7.4
httpx==0.23.0
jinja2==3.1.2
